all good

+++ src/main/java/pojos/DailyDataAnalysis.java
@@ -25,7 +25,7 @@ public class DailyDataAnalysis {
 	private Double	PDIDaily	;
 	private Double	MDIDaily	;
 	private String  MACDBuyOrSell;
-	private Double  MACDAboveSignal;
+	private int  MACDAboveSignal;
 	private String  RSIBuyOrSell;
 	private String  BBBuyOrSell;
 	private String  ADXBuyOrSell;
@@ -154,10 +154,10 @@ public class DailyDataAnalysis {
 	public void setMACDBuyOrSell(String mACDBuyOrSell) {
 		MACDBuyOrSell = mACDBuyOrSell;
 	}
-	public Double getMACDAboveSignal() {
+	public int getMACDAboveSignal() {
 		return MACDAboveSignal;
 	}
-	public void setMACDAboveSignal(Double mACDAboveSignal) {
+	public void setMACDAboveSignal(int mACDAboveSignal) {
 		MACDAboveSignal = mACDAboveSignal;
 	}
 	public String getRSIBuyOrSell() {
